refactor: mv feedback widget to own component
