+++ packages/portal/src/components/feedback/FeedbackWidget.vue
@@ -1,0 +1,32 @@
+<template>
+  <div v-if="feedbackEnabled">
+    <client-only>
+      <div
+        id="europeana-feedback-widget"
+        data-api-url="/_api/jira-service-desk/feedback"
+        data-faq-url="/faq"
+        :data-locale="$i18n.locale"
+        data-qa="feedback widget"
+      />
+      <script
+        type="module"
+        src="https://cdn.jsdelivr.net/npm/@europeana/feedback-widget@0.2.0/dist/europeana-feedback-widget.js"
+        integrity="sha384-u9uk9KOvOrG5X4yAGn1xMj1P5YmrJWanSqij7tvI4eZTjI/UQGK98lq3e/I6XfuP"
+      />
+      <link
+        rel="stylesheet"
+        href="https://cdn.jsdelivr.net/npm/@europeana/feedback-widget@0.2.0/dist/europeana-feedback-widget.css"
+      >
+    </client-only>
+  </div>
+</template>
+
+<script>
+  export default {
+    data() {
+      return {
+        feedbackEnabled: this.$features?.jiraServiceDeskFeedbackForm && !!this.$config?.app?.baseUrl
+      };
+    }
+  };
+</script>
