+++ packages/portal/tests/unit/components/page/PageFooter.spec.js
@@ -11,11 +11,6 @@ localVue.component('SmartLink', SmartLink);
 const factory = ({ mocks = {} } = {}) => shallowMount(PageFooter, {
   localVue,
   mocks: {
-    $config: { app: { baseUrl: 'https://www.example.eu' } },
-    $features: {},
-    $i18n: {
-      locale: 'en'
-    },
     $store: {
       getters: {
         'debug/settings': { enabled: false }
@@ -60,28 +55,6 @@ describe('components/PageFooter', () => {
       wrapper.vm.$store.getters['debug/settings'] = { enabled: true };
 
       expect(wrapper.vm.showDebugLinkGroup).toBe(true);
-    });
-  });
-
-  describe('FeedbackWidget', () => {
-    describe('when feedback toggle disabled', () => {
-      it('is not loaded', () => {
-        const wrapper = factory();
-
-        const feedbackWidget = wrapper.find('[data-qa="feedback widget"]');
-
-        expect(feedbackWidget.exists()).toBe(false);
-      });
-    });
-
-    describe('when feedback toggle enabled', () => {
-      it('is rendered', () => {
-        const wrapper = factory({ mocks: { $features: { jiraServiceDeskFeedbackForm: true } } });
-
-        const feedbackWidget = wrapper.find('[data-qa="feedback widget"]');
-
-        expect(feedbackWidget.exists()).toBe(true);
-      });
     });
   });
 });
